Revert to building for android-19.
android-21 introduces some changes to the system headers, making newer builds incompatible with pre-Android 5.0 systems. In order to maintain compatibility we have to compile with older headers, which will generate a lot more warnings, but will at least work. The resulting binaries seem to be working with Android 5.0+ for now, but that might change in time. So the build system should probably be updated in the future to fix this.